+++ index.js
@@ -1,9 +1,11 @@
 import GroundControl from "./ground-control.js";
 import InputControl from "./input-control.js";
 import ToggleControl from "./toggle-control.js";
+import ToggleSwitch from "./switch-control.js";
 
 export default {
   GroundControl,
   InputControl,
   ToggleControl,
+  ToggleSwitch,
 };