+++ input-control.js
@@ -20,6 +20,7 @@ export default class InputControl extends GroundControl {
 
   constructor() {
     super();
+    GroundControl.blockDisplay(this);
 
     this.addEventListener('change', this.onInputChange);
     this.addEventListener('input', this.onInputChange);