+++ toggle-control.js
@@ -33,6 +33,7 @@ export default class ToggleControl extends GroundControl {
 
   constructor() {
     super();
+    GroundControl.blockDisplay(this);
     this.addEventListener('click', this.onTogglePress);
   }